+++ src/main/java/com/libqa/web/repository/FeedActionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FeedActionRepository extends JpaRepository<FeedAction, Long> {
-    List<FeedAction> findByFeedIdAndUserId(long feedId, Integer userId);
-    List<FeedAction> findByFeedReplyIdAndUserId(long feedReplyId, Integer userId);
+public interface FeedActionRepository extends JpaRepository<FeedAction, Integer> {
+    List<FeedAction> findByFeedIdAndUserId(Integer feedId, Integer userId);
+    List<FeedAction> findByFeedReplyIdAndUserId(Integer feedReplyId, Integer userId);
 }
